refactor: оптимизирует функцию генерации фильтров
* Извлекает содержимое из папки global в папку src
