+++ src/main.js
@@ -1,6 +1,6 @@
-import DestinationsModel from './global/model/destinations-model.js';
-import TypeOffersModel from './global/model/type-offers-model.js';
-import EventsModel from './global/model/events-model.js';
+import DestinationsModel from './model/destinations-model.js';
+import TypeOffersModel from './model/type-offers-model.js';
+import EventsModel from './model/events-model.js';
 import TripMainPresenter from './modules/trip-main/presenter/trip-main-presenter.js';
 import EventsPresenter from './modules/events/presenter/events-presenter.js';
