+++ src/model/type-offers-model.js
